# user/api/openapi.yml
openapi: 3.0.0

servers:
  - description: Local environment
    url: 'https://localhost:8020'

info:
  title: User API
  version: 1.0.0
  description: |-
    This API allows you to manage users data, including creating, retrieving, updating, and deleting user records.
    
    Changelog:
    * 1.0.0 - Initial version of the API.  
      Provide CRUD operations and security features.

paths:
  /api/users:
    get:
      summary: Retrieve all users.
      description: Retrieve all users.
      tags:
        - Users
      operationId: getUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    post:
      summary: Create a user.
      description: Create a user.
      tags:
        - Users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The user was created successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: The unique identifier of the created user.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /api/users/{id}:
    get:
      summary: Retrieve the user from id.
      description: Retrieve the user from id.
      tags:
        - Users
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/idPath'
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '302':
            description: If the etag header is provided and the data has not changed, a 302 status code is returned to indicate that the client can use the cached version of the data.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    put:
      summary: Update a user from id.
      description: Update a user from id.
      tags:
        - Users
      operationId: updateUserById
      parameters:
        - $ref: '#/components/parameters/idPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '204':
          description: The user was updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    delete:
      summary: Delete a user from id.
      description: Delete a user from id.
      tags:
        - Users
      operationId: deleteUserById
      parameters:
        - $ref: '#/components/parameters/idPath'
      responses:
        '204':
          description: The user was deleted successfully
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'

components:
  parameters:
    idPath:
      name: id
      in: path
      description: The unique identifier of the user.
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    User:
      type: object
      properties:
        firstName:
          type: string
          description: The first name of the user.
          example: Chuck
        name:
          type: string
          description: The last name of the user.
          example: Norris
        emailId:
          type: integer
          format: int64
          description: The unique identifier of the email associated with the user.
          example: 1
        modified:
          type: localdatetime
          description: The local date and time when the user record was last modified.
          example: 2024-06-01T12:00:00.000000
    Error:
      type: object
      properties:
        code:
          type: string
          description: A code representing the error type.
          example: user.unexpected
        message:
          type: string
          description: A message describing the error.
          example: Unexpected error
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Unauthenticated user
    Unauthorized:
      description: The user is not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedError:
      description: An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

security:
  - basicAuth: []